+++ spec/verify.sh
@@ -4,10 +4,12 @@ if [[ "$OSTYPE" == "darwin"* ]]; then
     url="https://github.com/jedisct1/minisign/releases/download/0.11/minisign-0.11-macos.zip"
     curl -sL $url -o test/generated/minisign.zip
     unzip -o test/generated/minisign.zip -d test/generated
-    test/generated/minisign -Vm test/generated/*.txt -p test/minisign.pub
 else
     url="https://github.com/jedisct1/minisign/releases/download/0.11/minisign-0.11-linux.tar.gz"
     curl -sL $url -o test/generated/minisign.tar.gz
     tar -xvzf test/generated/minisign.tar.gz -C test/generated
-    test/generated/minisign-linux/x86_64/minisign -Vm test/generated/*.txt -p test/minisign.pub
+    mv test/generated/minisign-linux/x86_64/minisign test/generated/minisign
 fi
+
+test/generated/minisign -Vm test/generated/encrypted-key.txt -p test/minisign.pub || exit 1
+test/generated/minisign -Vm test/generated/unencrypted-key.txt -p test/unencrypted.pub || exit 1