+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Support signing with unencrypted keys
+
 ## [0.0.8] - 2024-02-03
 
 ### Added